+++ modules/admin/assets/Jquery.php
@@ -2,12 +2,6 @@
 
 namespace admin\assets;
 
-class Jquery extends \luya\web\Asset
+class Jquery extends \yii\web\JqueryAsset
 {
-    public $sourcePath = '@bower';
-
-    public $js = [
-        'jquery/dist/jquery.min.js',
-        'jquery-ui/jquery-ui.min.js',
-    ];
 }